pyproject.toml
[tool.poetry]
name = "horror_tv_schedule"
version = "0.1.0"
description = ""
authors = ["araki777 <arakichi777@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.10"
bs4 = "^0.0.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
selenium = "^4.11.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
